Compilation and 64-bit fixes to libbe.so sources.
Fixed the usual issues - printf format strings, uint32 instead of
addr_t, etc. One thing that isn't so nice is several places where
BList is used to store (u)int32, these require a double cast to addr_t
then void* to silence a warning on x86_64.

+++ headers/os/interface/TextView.h
@@ -411,7 +411,7 @@ private:
 			void				_ShowContextMenu(BPoint where);
 
 			void				_FilterDisallowedChars(char* text,
-									int32& length, text_run_array* runArray);
+									ssize_t& length, text_run_array* runArray);
 
 private:
 			BPrivate::TextGapBuffer*	fText;
